+++ examples/jxlinfo.c
@@ -106,6 +106,11 @@ int PrintBasicInfo(FILE* file) {
         printf("num_loops: %u\n", info.animation.num_loops);
         printf("have_timecodes: %d\n", info.animation.have_timecodes);
       }
+      printf("have_intrinsic_size: %d\n", info.have_intrinsic_size);
+      if (info.have_intrinsic_size) {
+        printf("intrinsic xsize: %u\n", info.intrinsic_size.xsize);
+        printf("intrinsic ysize: %u\n", info.intrinsic_size.ysize);
+      }
       const char* const orientation_string[8] = {
           "Normal",          "Flipped horizontally",
           "Upside down",     "Flipped vertically",